Moving MigrationName back to migrations_internals
